requirements.txt
tensorflow>=2.9
numpy
Pillow
matplotlib